Using a double to represent time in seconds instead of an int to represent time in ms. This gives us more width and depth

+++ src/state_manager.cpp
@@ -62,7 +62,7 @@ void StateManager::validate() {
   jassert(presets.isValid());
 }
 
-void StateManager::addClip(const juce::String& name, int start) {
+void StateManager::addClip(const juce::String& name, double start) {
   JUCE_ASSERT_MESSAGE_THREAD
 
   juce::ValueTree clip { ID::CLIP };
